Renamed file and module liststyleediting -> listpropertiesediting
