+++ packages/ckeditor5-list/src/index.js
@@ -11,7 +11,7 @@ export { default as List } from './list';
 export { default as ListEditing } from './listediting';
 export { default as ListUI } from './listui';
 export { default as ListStyle } from './liststyle';
-export { default as ListPropertiesEditing } from './liststyleediting';
+export { default as ListPropertiesEditing } from './listpropertiesediting';
 export { default as ListStyleUI } from './liststyleui';
 export { default as TodoList } from './todolist';
 export { default as TodoListEditing } from './todolistediting';
