+++ packages/ckeditor5-list/tests/liststylecommand.js
@@ -6,7 +6,7 @@
 import { setData, getData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
 import VirtualTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/virtualtesteditor';
 import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph';
-import ListPropertiesEditing from '../src/liststyleediting';
+import ListPropertiesEditing from '../src/listpropertiesediting';
 
 describe( 'ListStyleCommand', () => {
 	let editor, model, bulletedListCommand, numberedListCommand, listStyleCommand;
