+++ scripts/docs/plugins-collection.json
@@ -122,7 +122,7 @@
   "packages/ckeditor5-list/src/list.js": "List",
   "packages/ckeditor5-list/src/listediting.js": "Listediting",
   "packages/ckeditor5-list/src/liststyle.js": "Liststyle",
-  "packages/ckeditor5-list/src/liststyleediting.js": "Liststyleediting",
+  "packages/ckeditor5-list/src/listpropertiesediting.js": "Liststyleediting",
   "packages/ckeditor5-list/src/liststyleui.js": "Liststyleui",
   "packages/ckeditor5-list/src/listui.js": "Listui",
   "packages/ckeditor5-list/src/todolist.js": "Todolist",
